# swagger.yaml
swagger: "2.0"
info:
  title: Toohak API
  description: |
    API description in for Toohak.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will improve the usability of the frontend.
  version: 1.0.0
schemes:
  - http
...
components:
  prim:
    Error:
      type: string
      example: "Unknown Type: string - error"
...
    Token:
      type: string
      example: 23748
...
    QuizId:
      type: integer
      example: 5546
...
  body:
...
    QuizNewIter2:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          token:
            $ref: '#/components/prim/Token'
          name:
            type: string
            example: My Quiz Name
          description:
            type: string
            example: A description of my quiz
...
  return:
...
    Error:
      properties:
        error:
          type: string
            - error
          $ref: '#/components/prim/Error'
...
    QuizNew:
      type: object
      properties:
        quizId:
          $ref: '#/components/prim/QuizId'
...
paths:
...
  /v1/admin/quiz:
    post:
      summary: Create a new quiz
      description: Given basic details about a new quiz, create one for the logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/QuizNewIter2'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizNew'
        401:
          description: Token is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains invalid characters. Valid characters are alphanumeric and spaces
            - Name is either less than 3 characters long or more than 30 characters long
            - Name is already used by the current logged in user for another quiz
            - Description is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
...